@@ -12,10 +12,10 @@
   },
   "devDependencies": {
     "@types/node": "^14.14.35",
-    "@typescript-eslint/eslint-plugin": "^4.19.0",
-    "@typescript-eslint/parser": "^4.19.0",
+    "@typescript-eslint/eslint-plugin": "^5.22.0",
+    "@typescript-eslint/parser": "^5.22.0",
     "after-all": "^2.0.2",
-    "eslint": "^8.13.0",
+    "eslint": "^8.15.0",
     "standard": "^17.0.0",
     "tape": "^5.2.2",
     "ts-node-dev": "^1.1.8",
