@@ -4,7 +4,7 @@ const service = new Bonjour()
 
 service.publish({ name: 'My Web Server', type: 'http', port: 3000 })
 
-let browser = service.find({ type: 'http' })
+const browser = service.find({ type: 'http' })
 
 browser.on('up', (service) => {
   console.log('up', service)
