@@ -54,7 +54,7 @@ export class Bonjour {
      * @param callback Callback when device found
      * @returns
      */
-    public findOne(opts: BrowserConfig | undefined = undefined, timeout = 10000, callback: CallableFunction): Browser {
+    public findOne(opts: BrowserConfig | undefined = undefined, timeout = 10000, callback?: CallableFunction): Browser {
         const browser: Browser = new Browser(this.server.mdns, opts)
         var timer: any
         browser.once('up', (service: Service) => {
