@@ -9,17 +9,22 @@ const { Bonjour } = require('../dist')
 
 const getAddresses = function () {
   const addresses = []
-  const itrs = os.networkInterfaces()
-  for (const i in itrs) {
-    const addrs = itrs[i]
-    for (const j in addrs) {
-      if (addrs[j].internal === false) {
-        addresses.push(addrs[j].address)
+  const itrs = Object.values(os.networkInterfaces())
+  for (const addrs of itrs) {
+    for (const { internal, mac, address } of addrs) {
+      if (internal === false && mac !== '00:00:00:00:00:00' && !addresses.includes(address)) {
+        addresses.push(address)
       }
     }
   }
   return addresses
 }
+
+const filterDuplicates = (input) => input.reduce((prev, curr) => {
+  const obj = prev
+  if (!obj.includes(curr)) prev.push(curr)
+  return obj
+}, [])
 
 const port = function (cb) {
   const s = dgram.createSocket('udp4')
@@ -82,12 +87,12 @@ test('bonjour.find', function (bonjour, t) {
         t.equal(s.name, 'Foo Bar')
         t.equal(s.fqdn, 'Foo Bar._test._tcp.local')
         t.deepEqual(s.txt, {})
-        t.deepEqual(s.rawTxt, Buffer.from('00', 'hex'))
+        t.deepEqual(s.rawTxt, [])
       } else {
         t.equal(s.name, 'Baz')
         t.equal(s.fqdn, 'Baz._test._tcp.local')
         t.deepEqual(s.txt, { foo: 'bar' })
-        t.deepEqual(s.rawTxt, Buffer.from('07666f6f3d626172', 'hex'))
+        t.deepEqual(s.rawTxt, [Buffer.from('666f6f3d626172', 'hex')])
       }
       t.equal(s.host, os.hostname())
       t.equal(s.port, 3000)
@@ -98,7 +103,7 @@ test('bonjour.find', function (bonjour, t) {
       t.ok(Number.isFinite(s.referer.port))
       t.ok(Number.isFinite(s.referer.size))
       t.deepEqual(s.subtypes, [])
-      t.deepEqual(s.addresses.sort(), getAddresses().sort())
+      t.deepEqual(filterDuplicates(s.addresses.sort()), getAddresses().sort())
 
       if (++ups === 2) {
         // use timeout in an attempt to make sure the invalid record doesn't
@@ -122,8 +127,8 @@ test('bonjour.find - binary txt', function (bonjour, t) {
 
     browser.on('up', function (s) {
       t.equal(s.name, 'Foo')
-      t.deepEqual(s.txt, { bar: Buffer.from('buz') })
-      t.deepEqual(s.rawTxt, Buffer.from('076261723d62757a', 'hex'))
+      t.deepEqual(s.txt, { bar: 'buz' })
+      t.deepEqual(s.rawTxt, [Buffer.from('6261723d62757a', 'hex')])
       bonjour.destroy()
       t.end()
     })
@@ -153,13 +158,12 @@ test('bonjour.find - down event', function (bonjour, t) {
 
 test('bonjour.findOne - callback', function (bonjour, t) {
   const next = afterAll(function () {
-    bonjour.findOne({ type: 'test' }, function (s) {
+    bonjour.findOne({ type: 'test' }, undefined, function (s) {
       t.equal(s.name, 'Callback')
       bonjour.destroy()
       t.end()
     })
   })
-
   bonjour.publish({ name: 'Invalid', type: 'test2', port: 3000 }).on('up', next())
   bonjour.publish({ name: 'Callback', type: 'test', port: 3000 }).on('up', next())
 })
@@ -176,4 +180,34 @@ test('bonjour.findOne - emitter', function (bonjour, t) {
 
   bonjour.publish({ name: 'Emitter', type: 'test', port: 3000 }).on('up', next())
   bonjour.publish({ name: 'Invalid', type: 'test2', port: 3000 }).on('up', next())
+})
+
+test('bonjour.publish multiple', (bonjour, t) => {
+  let counter = 0
+  const onUp = () => {
+    counter += 1
+    if (counter === 3) {
+      bonjour.destroy()
+      t.end()
+    }
+  }
+  bonjour.publish({ name: 'A', type: 'A', port: 3000 }).on('up', onUp)
+  bonjour.publish({ name: 'B', type: 'B', port: 3000 }).on('up', onUp)
+  bonjour.publish({ name: 'C', type: 'C', port: 3000 }).on('up', onUp)
+})
+
+test('bonjour.publish multiple nested', (bonjour, t) => {
+  let counter = 0
+  bonjour.publish({ name: 'A', type: 'A', port: 3000 }).on('up', () => {
+    counter++
+    bonjour.publish({ name: 'B', type: 'B', port: 3000 }).on('up', () => {
+      counter++
+      bonjour.publish({ name: 'C', type: 'C', port: 3000 }).on('up', () => {
+        counter++
+        t.equal(counter, 3)
+        bonjour.destroy()
+        t.end()
+      })
+    })
+  })
 })
