@@ -6,26 +6,25 @@ const { Service } = require('../dist/lib/service')
 
 const getAddressesRecords = function (host) {
   const records = []
-  const itrs = os.networkInterfaces()
-  for (const i in itrs) {
-    const addrs = itrs[i]
-    for (const j in addrs) {
-      if (addrs[j].internal === false) {
-        records.push({ data: addrs[j].address, name: host, ttl: 120, type: addrs[j].family === 'IPv4' ? 'A' : 'AAAA' })
+  const itrs = Object.values(os.networkInterfaces())
+  for (const addrs of itrs) {
+    for (const { internal, address, family, mac } of addrs) {
+      if (internal === false && mac !== '00:00:00:00:00:00') {
+        records.push({ data: address, name: host, ttl: 120, type: family === 'IPv4' ? 'A' : 'AAAA' })
       }
     }
   }
   return records
 }
 
-test('no name', function (t) {
+test('service must throw an exception if no name is specified', function (t) {
   t.throws(function () {
     new Service({ type: 'http', port: 3000 }) // eslint-disable-line no-new
   }, 'Required name not given')
   t.end()
 })
 
-test('no type', function (t) {
+test('service must throw an exception if no type is specified', function (t) {
   t.throws(function () {
     new Service({ name: 'Foo Bar', port: 3000 }) // eslint-disable-line no-new
   }, 'Required type not given')
@@ -47,8 +46,8 @@ test('minimal', function (t) {
   t.equal(s.host, os.hostname())
   t.equal(s.port, 3000)
   t.equal(s.fqdn, 'Foo Bar._http._tcp.local')
-  t.equal(s.txt, null)
-  t.equal(s.subtypes, null)
+  t.equal(s.txt, undefined)
+  t.equal(s.subtypes, undefined)
   t.equal(s.published, false)
   t.end()
 })
@@ -76,7 +75,7 @@ test('_records() - minimal', function (t) {
   t.deepEqual(s.records(), [
     { data: s.fqdn, name: '_http._tcp.local', ttl: 28800, type: 'PTR' },
     { data: { port: 3000, target: os.hostname() }, name: s.fqdn, ttl: 120, type: 'SRV' },
-    { data: Buffer.from('00', 'hex'), name: s.fqdn, ttl: 4500, type: 'TXT' }
+    { data: [], name: s.fqdn, ttl: 4500, type: 'TXT' }
   ].concat(getAddressesRecords(s.host)))
   t.end()
 })
@@ -86,7 +85,7 @@ test('_records() - everything', function (t) {
   t.deepEqual(s.records(), [
     { data: s.fqdn, name: '_http._tcp.local', ttl: 28800, type: 'PTR' },
     { data: { port: 3000, target: 'example.com' }, name: s.fqdn, ttl: 120, type: 'SRV' },
-    { data: Buffer.from('07666f6f3d626172', 'hex'), name: s.fqdn, ttl: 4500, type: 'TXT' }
+    { data: [Buffer.from('666f6f3d626172', 'hex')], name: s.fqdn, ttl: 4500, type: 'TXT' }
   ].concat(getAddressesRecords(s.host)))
   t.end()
 })
